+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Next
 
-### Added
+## [2.4.0] - 2025-07-20
 
 - `searchcount` component
 - `selectioncount` component